Retiros View Insert Ready

+++ src/main/java/co/edu/usbcali/demo/delegado/DelegadoDeNegocio.java
@@ -194,7 +194,7 @@ public class DelegadoDeNegocio implements IDelegadoDeNegocio {
 	}
 	
 	@Override
-	public int consultarUltimoRetiro() {
+	public Long consultarUltimoRetiro() {
 		return retirosLogica.consultarUltimo();
 	}
